requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
google-generativeai>=0.8,<1.0
pydantic>=2.0,<3.0